requirements.txt
agno==2.5.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
ddgs==9.10.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
fake-useragent==2.2.0
fastapi==0.133.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
groq==1.0.0
grpcio==1.78.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
numpy==2.4.2
openai==2.24.0
packaging==26.0
portalocker==3.2.0
postgrest==2.28.0
primp==1.1.1
propcache==0.4.1
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.0
realtime==2.28.0
requests==2.32.5
rich==14.3.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
